% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wos_get_records.R
\name{wos_get_records}
\alias{wos_get_records}
\title{Download references that match a given WOS query}
\usage{
wos_get_records(query, database = "WOS", limit = NULL, sleep = 1)
}
\arguments{
\item{query}{a \code{character} of length 1. The query to send to the WOS Starter
API. Visit the WOS documentation at:
\url{https://images.webofknowledge.com/images/help/WOK/contents.html} on
how to write a WOS query.}

\item{database}{a \code{character} of length 1. One among \code{BCI} (BIOSIS Citation
Index), \code{BIOABS} (Biological Abstracts), \code{BIOSIS} (BIOSIS Previews),
\code{CCC} (Current Contents Connect), \code{DIIDW} (Derwent Innovations Index),
\code{DRCI } (Data Citation Index), \code{MEDLINE} (Medline), \code{PPRN} (Preprint
Citation Index), \code{WOS} (Web of Science Core Collection), \code{ZOOREC}
(Zoological Record), and \code{WOK} (all databases).
Default is \code{WOS} (all databases).}

\item{limit}{an \code{numeric} of length 1. The number of records to retrieve.
Default is \code{NULL} (all possible records will be retrieved).}

\item{sleep}{an \code{numeric} of length 1. To not stress the WOS STARTER API, a
random number between 0 and \code{sleep} will be picked to suspend queries.}
}
\value{
A \code{data.frame} with \code{n} rows (where \code{n} is the total number of
references) and the following 21 variables:
\code{uid}: the unique identifier of the reference in the Web of Science system;
\code{document_type}: the document type;
\code{title}: the title of the reference;
\code{authors}: the authors of the reference;
\code{published_year}: the published year;
\code{published_month}: the published month;
\code{source}: the title of the source (journal, book, etc.) in which the
reference was published;
\code{volume}: the volume number;
\code{issue}: the issue number;
\code{pages}: the pages range in the source;
\code{no_article}: the article number;
\code{supplement_number}: the supplement number (if applicable);
\code{special_issue}: \code{SI} in case of a special issue (if applicable);
\code{book_editors}: the book authors (if applicable);
\code{keywords}: the authors keywords;
\code{doi}: the Digital Object Identifier;
\code{eissn}: the Electronic International Standard Identifier Number;
\code{issn}: the International Standard Identifier Number;
\code{isbn}: International Standard Book Number.
\code{pmid}: PubMed identifier.
\code{citations}: the number of citations in the database.
}
\description{
This function sends a query to the Web Of Science Starter API
(\url{https://developer.clarivate.com/apis/wos-starter}) and returns
references that match this query.

To learn how to write a WOS query, users can read the WOS documentation
available at:
\url{https://images.webofknowledge.com/images/help/WOK/contents.html}.
A list of WOS field tags is available at:
\url{https://images.webofknowledge.com/images/help/WOS/hs_wos_fieldtags.html}.

It's strongly recommended to use the function \link{wos_search} before to
have an idea on how many records you will download.
}
\examples{
\dontrun{
## Write query to retrieve references of one author ----
query <- "AU=(\"Casajus Nicolas\")"

## Check the number of records ----
wos_search(query)

## Download metadata of records ----
refs <- wos_get_records(query)
}
}
